refactor(holochain_state): split dht_store tests into submodule; doc cleanup
- Convert dht_store.rs → dht_store/mod.rs + dht_store/tests.rs.
- Drop block-comment section headers from tests.
- Remove SysOutcome::AwaitingDeps and AppOutcome::AwaitingDeps (status
  stays NULL for awaiting-deps; no update needed).
- Split record_sys_validation_outcome into record_chain_op_sys_validation_outcome
  and record_warrant_sys_validation_outcome.
- Remove private-entry branch from record_incoming_ops (network ops never
  carry private entries).
- Rename apply_countersigning_success → clear_op_withhold_publish.
- Rename action_hash_to_entry_hash → entry_hash_from_chain_op_action.
- Drop abandoned_at branch from compute_chain_op_validation_status.
- Use OpValidity in reject_chain_op (force_reject_limbo_chain_op for limbo).
- DhtStoreError::ChainOpPublishMissing: remove the hash payload.
- Rewrite doc comments to describe code, not transitional context.
